python_rag/requirements.txt
openai>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
streamlit>=1.30.0
python-dotenv>=1.0.0